renamed benchmarking to CLI
