+++ cli/auto_metrics/AutoMetric.py
