+++ cli/auto_metrics/CellCountMetric.py
