+++ cli/auto_metrics/CellTypingMetric.py
