+++ cli/core/ollama_wrapper.py
