+++ cli/prompt_testing/MultiAgentTester.py
@@ -1,7 +1,21 @@
 #!/usr/bin/env python3
 """
-Interactive Auto Agent System Tester (v1.2-auto)
-==========================================
+Interactive Agent System Tester (v1.1)
+======================================
+• **New in v1.1** – Smarter delegation detection.
+  The router now recognises any of the following patterns in an assistant reply
+  when deciding to switch agents:
+
+  ```text
+  //delegate_to_coder
+  delegate_to_coder
+  `delegate_to_coder`
+  Executing command: `delegate_to_coder`
+  ```
+
+  No need to rigidly start the reply with the token – the regex scans the whole
+  message. Once detected, we alert the user ("🔄 Routing to …") and prepend the
+  new agent’s system prompt.
 """
 from __future__ import annotations
 
@@ -16,10 +30,9 @@ import time
 from datetime import datetime
 from pathlib import Path
 from typing import List, Tuple, Optional, Dict
-
 from rich.table import Table
-# -- Pick LLM backend ---------------------------------------------------
 from rich.prompt import Prompt
+
 BACKEND_CHOICE = Prompt.ask(
     "LLM backend",
     choices=["chatgpt", "ollama"],
@@ -35,7 +48,7 @@ if BACKEND_CHOICE == "ollama":
 try:
     from dotenv import load_dotenv
     if BACKEND_CHOICE == "ollama":
-        from benchmarking.core.ollama_wrapper import OllamaClient as OpenAI
+        from cli.core.ollama_wrapper import OllamaClient as OpenAI
         APIError = Exception  # Ollama does not have a specific APIError
     else:
         from openai import OpenAI, APIError
@@ -44,16 +57,15 @@ try:
 except ImportError as e:
     print(f"Missing dependency: {e}", file=sys.stderr)
     sys.exit(1)
-
 # ── Agent framework ---------------------------------------------------------
 try:
-    from benchmarking.agents.AgentSystem import AgentSystem, Agent
+    from cli.agents.AgentSystem import AgentSystem, Agent
 except ImportError:
     print("[ERROR] Could not import backend.agents.agent_system", file=sys.stderr)
     raise
 
 # ── Local helpers -----------------------------------------------------------
-from benchmarking.core.io_helpers import (
+from cli.core.io_helpers import (
     extract_python_code,
     display,
     select_dataset,
@@ -62,7 +74,7 @@ from benchmarking.core.io_helpers import (
     format_execute_response,
     load_bp_json
 )
-from benchmarking.core.sandbox_management import (
+from cli.core.sandbox_management import (
     init_docker,
     init_singularity,
     init_singularity_exec,
@@ -78,73 +90,26 @@ ENV_FILE = PARENT_DIR / ".env"
 SANDBOX_DATA_PATH = "/workspace/dataset.h5ad"
 SANDBOX_RESOURCES_DIR = "/workspace/resources"
 
-# ── Benchmark persistence --------------------------------------------------
-timestamp = datetime.utcnow().strftime("%Y%m%d-%H%M%S")
-_LEDGER_PATH = OUTPUTS_DIR / f"benchmark_history_{timestamp}.jsonl"
-_SNIPPET_DIR = OUTPUTS_DIR / "snippets"
-_SNIPPET_DIR.mkdir(exist_ok=True, parents=True)
-_LEDGER_PATH.parent.mkdir(exist_ok=True, parents=True)
-
-def _dump_code_snippet(run_id: str, code: str) -> str:
-    """
-    Write <run_id>.py under outputs/snippets/ and return the relative path.
-    """
-    snippet_path = _SNIPPET_DIR / f"{run_id}.py"
-    snippet_path.write_text(code, encoding="utf-8")
-    return str(snippet_path.relative_to(OUTPUTS_DIR))
-
-def _save_benchmark_record(*, run_id: str, results: dict, meta: dict, code: str | None):
-    """
-    Append a JSONL record containing timestamp, dataset metadata, metrics, and
-    a pointer to (or inline copy of) the integration code.
-    """
-    record = {
-        "ts": datetime.utcnow().isoformat(timespec="seconds") + "Z",
-        "run": run_id,
-        "dataset": meta.get("name"),
-        "results": results,
-    }
-    if code:
-        record["code_path"] = _dump_code_snippet(run_id, code)
-    with _LEDGER_PATH.open("a") as fh:
-        fh.write(json.dumps(record) + "\n")
-
 # ===========================================================================
 # 1 · Backend selection
 # ===========================================================================
-backend = Prompt.ask(
-    "Choose sandbox backend", choices=["docker", "singularity", "singularity-exec"], default="docker"
-)
-force_refresh = (
-    Prompt.ask("Force refresh environment?", choices=["y", "n"], default="n").lower() == "y"
-)
+backend = Prompt.ask("Choose sandbox backend", choices=["docker", "singularity", "singularity-exec"], default="docker")
+force_refresh = Prompt.ask("Force refresh environment?", choices=["y", "n"], default="n").lower() == "y"
 is_exec_mode = backend == "singularity-exec"
 
 if backend == "docker":
-    (
-        _BackendManager,
-        _SANDBOX_HANDLE,
-        COPY_CMD,
-        EXECUTE_ENDPOINT,
-        STATUS_ENDPOINT,
-    ) = init_docker(SCRIPT_DIR, subprocess, console, force_refresh)
+    _BackendManager, _SANDBOX_HANDLE, COPY_CMD, EXECUTE_ENDPOINT, STATUS_ENDPOINT = init_docker(
+        SCRIPT_DIR, subprocess, console, force_refresh
+    )
     SANDBOX_DATA_PATH = "dataset.h5ad"
 elif backend == "singularity":
-    (
-        _BackendManager,
-        _SANDBOX_HANDLE,
-        COPY_CMD,
-        EXECUTE_ENDPOINT,
-        STATUS_ENDPOINT,
-    ) = init_singularity(SCRIPT_DIR, subprocess, console, force_refresh)
+    _BackendManager, _SANDBOX_HANDLE, COPY_CMD, EXECUTE_ENDPOINT, STATUS_ENDPOINT = init_singularity(
+        SCRIPT_DIR, subprocess, console, force_refresh
+    )
 elif backend == "singularity-exec":
-    (
-        _BackendManager,
-        _SANDBOX_HANDLE,
-        COPY_CMD,
-        EXECUTE_ENDPOINT,
-        STATUS_ENDPOINT,
-    ) = init_singularity_exec(SCRIPT_DIR, SANDBOX_DATA_PATH, subprocess, console, force_refresh)
+    _BackendManager, _SANDBOX_HANDLE, COPY_CMD, EXECUTE_ENDPOINT, STATUS_ENDPOINT = init_singularity_exec(
+        SCRIPT_DIR, SANDBOX_DATA_PATH, subprocess, console, force_refresh
+    )
 else:
     console.print("[red]Unknown backend.")
     sys.exit(1)
@@ -152,18 +117,17 @@ else:
 # ===========================================================================
 # 2 · Agent helpers
 # ===========================================================================
+
 def load_agent_system() -> Tuple[AgentSystem, Agent, str]:
-    """Load the agent system from a JSON blueprint."""
     bp = load_bp_json(console)
-    if not bp.exists():
-        console.print(f"[red]Blueprint {bp} not found.")
-        sys.exit(1)
     system = AgentSystem.load_from_json(str(bp))
     driver_name = Prompt.ask("Driver agent", choices=list(system.agents.keys()), default=list(system.agents)[0])
     driver = system.get_agent(driver_name)
     instr = system.get_instructions()
     return system, driver, instr
 
+# Smarter regex – matches inline/backtick/explicit styles
+# Match variations like //<backtick>delegate_to_coder<backtick>, with optional punctuation.
 _DELEG_RE = re.compile(r"delegate_to_([A-Za-z0-9_]+)")
 
 def detect_delegation(msg: str) -> Optional[str]:
@@ -171,8 +135,8 @@ def detect_delegation(msg: str) -> Optional[str]:
     m = _DELEG_RE.search(msg)
     return f"delegate_to_{m.group(1)}" if m else None
 
+
 def api_alive(url: str, tries: int = 10) -> bool:
-    """Check if the API is responsive."""
     if is_exec_mode:
         return True
     for _ in range(tries):
@@ -184,22 +148,10 @@ def api_alive(url: str, tries: int = 10) -> bool:
     return False
 
 # ===========================================================================
-# 3 · Interactive *or* Automated loop
+# 3 · Interactive loop
 # ===========================================================================
-def run(
-    agent_system: AgentSystem,
-    agent: Agent,
-    roster_instr: str,
-    dataset: Path,
-    metadata: dict,
-    resources: List[Tuple[Path, str]],
-    benchmark_module: Optional[Path] = None,
-    *,
-    initial_user_message: str,
-    tries: int = 0,
-):
-    """Main driver"""
-    last_code_snippet: str | None = None  
+
+def run(agent_system: AgentSystem, agent: Agent, roster_instr: str, dataset: Path, metadata: dict, resources: List[Tuple[Path, str]], benchmark_modules: Optional[list[Path]] = None):
     mgr = _BackendManager()
     console.print(f"Launching sandbox ({backend})…")
 
@@ -226,9 +178,10 @@ def run(
         return roster_instr + "\n\n" + a.get_full_prompt(agent_system.global_policy) + "\n\n" + analysis_ctx
 
     history = [{"role": "system", "content": build_system(agent)}]
-    history.append({"role": "user", "content": initial_user_message})
+    first_user = "Beginning interactive session. You can ask questions or give commands."
+    history.append({"role": "user", "content": first_user})
     display(console, "system", history[0]["content"])
-    display(console, "user", initial_user_message)
+    display(console, "user", first_user)
 
     if BACKEND_CHOICE == "chatgpt":
         if not os.getenv("OPENAI_API_KEY"):
@@ -240,8 +193,6 @@ def run(
         openai = OpenAI(host=OLLAMA_HOST, model="deepseek-r1:70b")
     current_agent = agent
     turn = 0
-
-    tries_left = tries
 
     while True:
         turn += 1
@@ -255,7 +206,6 @@ def run(
         history.append({"role": "assistant", "content": msg})
         display(console, f"assistant ({current_agent.name})", msg)
 
-        # ── Delegation --------------------------------------------------------
         cmd = detect_delegation(msg)
         if cmd and cmd in current_agent.commands:
             tgt = current_agent.commands[cmd].target_agent
@@ -263,7 +213,7 @@ def run(
             if new_agent:
                 console.print(f"[yellow]🔄 Routing to '{tgt}' via {cmd}")
                 history.append({"role": "assistant", "content": f"🔄 Routing to **{tgt}** (command `{cmd}`)"})
-                
+                  
                 # INJECT LOADED CODE SAMPLES ON DELEGATION ---
                 if new_agent.code_samples:
                     sample_context = "Here are some relevant code samples for your task:"
@@ -278,10 +228,8 @@ def run(
                 history.insert(0, {"role": "system", "content": build_system(new_agent)})
                 continue
 
-        # ── Inline code execution -------------------------------------------
         code = extract_python_code(msg)
         if code:
-            last_code_snippet = code   
             console.print("[cyan]Executing code…[/cyan]")
             try:
                 if is_exec_mode:
@@ -296,74 +244,99 @@ def run(
 
             history.append({"role": "user", "content": feedback})
             display(console, "user", feedback)
-
-        # ── Automatic benchmarking (v1.2 addition) --------------------------
-        if benchmark_module:
-            result_str = run_benchmark(mgr, benchmark_module, metadata, current_agent.name, last_code_snippet)
-            if result_str:
-                history.append({"role": "user", "content": result_str})
-                display(console, "user", result_str)
-        tries_left -= 1
-        if tries_left <= 0:
+            
+        def input_loop():
+            if benchmark_modules:
+                console.print("\n[bold]Next message (blank = continue, 'benchmark' to run benchmarks, 'exit' to quit):[/bold]")
+            else:
+                console.print("\n[bold]Next message (blank = continue, 'exit' to quit):[/bold]")
+            try:
+                user_in = input().strip()
+            except (EOFError, KeyboardInterrupt):
+                user_in = "exit"
+            if user_in.lower() in {"exit", "quit"}:
+                return "break"
+            if user_in.lower() == "benchmark" and benchmark_modules:
+                for benchmark_module in benchmark_modules:
+                    run_benchmark(mgr, benchmark_module)
+                input_loop()  # Recurse to continue the loop after benchmarks
+            if user_in:
+                history.append({"role": "user", "content": user_in})
+                display(console, "user", user_in)
+        input_val = input_loop()
+        if input_val == "break":  # User chose to exit
             break
-        # Simulate blank *continue* from the user
-        history.append({"role": "user", "content": ""})
-        continue
+
     console.print("Stopping sandbox…")
     mgr.stop_container()
 
+
 # ===========================================================================
-# 4 · Benchmarking helpers (modified to *return* results)
+# 4 · Benchmarking
 # ===========================================================================
-def get_benchmark_module(console: Console, parent_dir: Path) -> Optional[Path]:
-    """Prompt user to select a benchmark module."""
+
+def get_benchmark_modules(console: Console, parent_dir: Path) -> Optional[list[Path]]:
+    """
+    Prompts the user to select a benchmark module from the available ones.
+    Returns the path to the selected module or None if no selection is made.
+    """
     benchmark_dir = parent_dir / "auto_metrics"
     if not benchmark_dir.exists():
         console.print("[red]No benchmarks directory found.[/red]")
         return None
 
-    modules = [m for m in benchmark_dir.glob("*.py") if m.name != "AutoMetric.py"]
-    if not modules:
+    module_names = list(benchmark_dir.glob("*.py"))
+    # remove AutoMetric.py from modules (it is the base class)
+    module_names = [m for m in module_names if m.name != "AutoMetric.py"]
+    if not module_names:
         console.print("[red]No benchmark modules found.[/red]")
         return None
 
     console.print("\n[bold]Available benchmark modules:[/bold]")
-    for i, mod in enumerate(modules, start=1):
+    for i, mod in enumerate(module_names, start=1):
         console.print(f"{i}. {mod.name}")
-
-    choice = Prompt.ask("Select a benchmark module by number (or press Enter to skip)", default="")
-    if not choice:
+    console.print(f"{len(module_names)+1}. Select All")
+    choices = Prompt.ask("Select benchmark modules by number  (e.g. 1 2 3 or 1,2,3) (or press Enter to skip)", default="")
+    choices = re.split(r'[,\s]+', choices) #User input must be seperated by commas or spaces 
+    
+    if not choices or choices == ['']:
         return None
 
-    try:
-        index = int(choice) - 1
-        if 0 <= index < len(modules):
-            return modules[index]
-        else:
-            console.print("[red]Invalid selection.[/red]")
+    modules = []
+    for choice in choices:
+        try: 
+            index = int(choice) - 1
+            if index == len(module_names): #Handles select all case 
+                return module_names
+            elif 0 <= index < len(module_names): 
+                modules.append(module_names[index])
+            else:
+                console.print("[red]Invalid selection.[/red]")
+                return None
+        except ValueError:
+            console.print("[red]Invalid input. Please enter a number.[/red]")
             return None
-    except ValueError:
-        console.print("[red]Invalid input. Please enter a number.[/red]")
-        return None
-
-
-def run_benchmark(mgr, benchmark_module: Path, metadata: dict,
-                  agent_name: str, code_snippet: str | None) -> str:
-    """Execute benchmark module and *return* a compact JSON string."""
-    console.print(f"\n[bold cyan]Running benchmark module: {benchmark_module.name}[/bold cyan]")
+    return modules 
+    
+def run_benchmark(mgr, benchmark_module: str):
+    """
+    Runs the benchmark module and displays the results.
+    """
+    console.print(f"\n[bold cyan]Running benchmark module: {benchmark_module}[/bold cyan]")
     autometric_base_path = benchmark_module.parent / "AutoMetric.py"
     try:
+        # Read the abstract base class definition
         with open(autometric_base_path, "r") as f:
             autometric_code = f.read()
+
         with open(benchmark_module, "r") as f:
             benchmark_code = f.read()
     except FileNotFoundError:
-        err = f"Benchmark module not found at: {benchmark_module}"
-        console.print(f"[red]{err}[/red]")
-        return err
+        console.print(f"[red]Benchmark module not found at: {benchmark_module}[/red]")
+        return
 
     code_to_execute = f"""
-# --- Code from AutoMetric.py ---
+# --- Code from AutoMetric.py --- 
 {autometric_code}
 # --- Code from {benchmark_module.name} ---
 {benchmark_code}
@@ -377,12 +350,15 @@ def run_benchmark(mgr, benchmark_module: Path, metadata: dict,
                 EXECUTE_ENDPOINT, json={"code": code_to_execute, "timeout": 300}, timeout=310
             ).json()
 
+        # Create a table to display the results
         table = Table(title="Benchmark Results")
         table.add_column("Metric", style="cyan")
         table.add_column("Value", style="magenta")
+
+        # Assuming the benchmark module returns a dictionary of results
         stdout = exec_result.get("stdout", "")
         try:
-            result_dict = json.loads(stdout.strip().splitlines()[-1])
+            result_dict = json.loads(stdout.strip().splitlines()[-1])  # Parse last printed line
         except Exception as e:
             console.print(f"[yellow]Warning: Could not parse JSON from stdout: {e}[/yellow]")
             result_dict = {}
@@ -390,57 +366,30 @@ def run_benchmark(mgr, benchmark_module: Path, metadata: dict,
         if exec_result.get("status") == "ok" and isinstance(result_dict, dict):
             for key, value in result_dict.items():
                 table.add_row(str(key), str(value))
-            _save_benchmark_record(
-                run_id=f"{benchmark_module.stem}:{agent_name}:{int(time.time())}",
-                results=result_dict,
-                meta=metadata,
-                code=code_snippet,
-            )
         else:
             table.add_row("Error", exec_result.get("stderr") or "An unknown error occurred.")
+
         console.print(table)
-        return "Benchmark results:\n" + json.dumps(result_dict or {"error": "see console"})
+
     except Exception as exc:
-        err_msg = f"Benchmark execution error: {exc}"
-        console.print(f"[red]{err_msg}[/red]")
-        return err_msg
+        console.print(f"[red]Benchmark execution error: {exc}[/red]")
 
 # ===========================================================================
-# 5 · Entry point (collect *tries* & initial message)
+# 4 · Entry point
 # ===========================================================================
+
 def main():
     load_dotenv(ENV_FILE)
     if not os.getenv("OPENAI_API_KEY"):
         console.print("[red]OPENAI_API_KEY not set in .env")
         sys.exit(1)
 
-    sys_, drv, roster = load_agent_system()
+    sys, drv, roster = load_agent_system()
     dp, meta = select_dataset(console, DATASETS_DIR)
-    benchmark_module = get_benchmark_module(console, PARENT_DIR)
+    benchmark_modules = get_benchmark_modules(console, PARENT_DIR)
     res = collect_resources(console, SANDBOX_RESOURCES_DIR)
+    run(sys, drv, roster, dp, meta, res, benchmark_modules)
 
-    initial_user_message = Prompt.ask(
-        "Initial user message", default="What should I do with this dataset?"
-    )
-    try:
-        tries = int(Prompt.ask("Number of automatic tries", default="1"))
-        if tries < 0:
-            raise ValueError
-    except ValueError:
-        console.print("[yellow]Invalid number – defaulting to 1.[/yellow]")
-        tries = 1
-
-    run(
-        sys_,
-        drv,
-        roster,
-        dp,
-        meta,
-        res,
-        benchmark_module,
-        initial_user_message=initial_user_message,
-        tries=tries,
-    )
 
 if __name__ == "__main__":
     try:
