+++ cli/sample_prompt_library/Basic_scRNA_Agent.txt
